Redo library tabs
