+++ src/components/molecules/mininav.module.scss
@@ -22,16 +22,17 @@
 	display: flex;
 	align-items: center;
 	flex-grow: 1;
-	flex-wrap: wrap;
-	margin-top: 8px;
 	margin-bottom: 8px;
 	> * {
-		margin-right: 24px;
+		margin-right: 16px;
 		margin-top: 8px;
 		margin-bottom: 8px;
 		padding-top: 3px; // Compensate for border-bottom
 		padding-bottom: 2px;
 		border-bottom: 1px solid transparent;
+		@media (min-width: $breakpoint-lg) {
+			margin-right: 24px;
+		}
 	}
 	a:hover {
 		text-decoration: none;
