+++ src/components/organisms/aboutNav.module.scss
@@ -1,3 +1,6 @@
 .wrapper {
+	margin-top: -8px;
 	margin-bottom: 24px;
+	overflow-x: scroll;
+	white-space: nowrap;
 }
