+++ src/containers/library/base.module.scss
@@ -1,8 +1,8 @@
 @import '../../styles/common';
 
 .wrapper {
-	margin-top: -16px;
+	margin-top: -8px;
 	@media (min-width: $breakpoint-lg) {
-		margin-top: -8px;
+		margin-top: 0;
 	}
 }
