--- src/containers/library/index.module.scss
@@ -1,5 +1,0 @@
-.seeAll {
-	display: flex;
-	justify-content: flex-end;
-	margin-bottom: 16px;
-}
