+++ src/containers/library/playlist/list.module.scss
@@ -1,8 +1,5 @@
-@import '../../styles/common';
+@import '../../../styles/common';
 
-.playlistsAlert {
-	margin-top: 24px;
-}
 .playlistsAlertDismiss {
 	margin-top: 16px;
 	font-size: 12px;
