+++ src/pages/[language]/library/playlists.ts
@@ -1,19 +1,19 @@
 import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 
-import LibraryCollections, {
-	ILibraryCollectionsProps,
-} from '@containers/library/collections';
+import LibraryPlaylists, {
+	ILibraryPlaylistsProps,
+} from '@containers/library/playlist/list';
 import { storeRequest } from '@lib/api';
 import { getLanguageIdByRoute } from '@lib/getLanguageIdByRoute';
 
-export default LibraryCollections;
+export default LibraryPlaylists;
 
 export function getServerSideProps({
 	req,
 	params,
 }: GetServerSidePropsContext<{
 	language: string;
-}>): GetServerSidePropsResult<ILibraryCollectionsProps> {
+}>): GetServerSidePropsResult<ILibraryPlaylistsProps> {
 	storeRequest(req);
 	const language = getLanguageIdByRoute(params?.language);
